ENH/FIX: Make the packet_generator on an XTCE definition generic
Currently, it is fixed to use a ccsds generator as the underlying
mechanism to yield packets. The XTCE definition itself can be used
as the generator and packets can be yielded based on the data parsed
during each iteration of the parsing. This means that someone
doesn't have to rely on the CCSDS specification or bring their own
generator to produce bytes, but can rather just utilize the XTCE
file they've already created.